name space is renamed to com.mierafacademy...
