+++ src/main/java/com/mierafacademy/studentreportcard/security/UserRepository.java
@@ -1,0 +1,12 @@
+package com.mierafacademy.studentreportcard.security;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import javax.transaction.Transactional;
+@Repository
+@Transactional
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findByEmail(String email);
+
+}
