+++ src/main/java/com/mierafacademy/studentreportcard/security/UserService.java
@@ -1,0 +1,8 @@
+package com.mierafacademy.studentreportcard.security;
+
+import org.springframework.security.core.userdetails.UserDetailsService;
+public interface UserService extends UserDetailsService {
+    User findByEmail(String email);
+
+    User save(UserRegistrationDto registration);
+}
