--- src/main/java/dev/cs/studentreportcard/security/UserRepository.java
@@ -1,11 +1,0 @@
-package dev.cs.studentreportcard.security;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import javax.transaction.Transactional;
-@Repository
-@Transactional
-public interface UserRepository extends JpaRepository<User, Long> {
-    User findByEmail(String email);
-
-}
