--- src/main/java/dev/cs/studentreportcard/security/UserService.java
@@ -1,6 +1,0 @@
-package dev.cs.studentreportcard.security;
-import org.springframework.security.core.userdetails.UserDetailsService;
-public interface UserService extends UserDetailsService {
-    User findByEmail(String email);
-    User save(UserRegistrationDto registration);
-}
