--- src/main/java/dev/cs/studentreportcard/services/CSVDataLoadingService.java
@@ -1,164 +1,0 @@
-package dev.cs.studentreportcard.services;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.opencsv.CSVReader;
-import dev.cs.studentreportcard.models.Student;
-import dev.cs.studentreportcard.models.StudentRecord;
-import dev.cs.studentreportcard.repositories.StudentRecordRepository;
-import dev.cs.studentreportcard.repositories.StudentRepository;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.stereotype.Service;
-
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.time.LocalDate;
-import java.util.List;
-
-@Service
-public class CSVDataLoadingService {
-    private final StudentRepository studentRepository;
-    private final StudentRecordRepository studentRecordRepository;
-
-    public CSVDataLoadingService(StudentRepository studentRepository, StudentRecordRepository studentRecordRepository) {
-        this.studentRepository       = studentRepository;
-        this.studentRecordRepository = studentRecordRepository;
-    }
-
-    public void loadStudentRecordDataFile() {
-        System.out.println("loadcsvdata method is called ");
-        try {
-            // ClassPathResource refers that everything under resource files will be included in the jar file of this project
-            // CSVReader reader = new CSVReader(new InputStreamReader(new ClassPathResource("studentrecorddata.csv").getInputStream()));
-            CSVReader reader = new CSVReader(new InputStreamReader(new ClassPathResource("studentrecorddata.csv").getInputStream()));
-            //TODO new File("C:/path/to/studentdata.csv"); if we need to upload from other location
-            String[] line;
-            boolean isHeader = true;
-            while ((line = reader.readNext()) != null) {
-                if (isHeader) {
-                    isHeader = false;
-                    continue;
-                }
-                // Optionally, also skip lines where all fields are blank eg. the last line w/c is blank
-                boolean allBlank = true;
-                for (String field : line) {
-                    if (field != null && !field.trim().isEmpty()) {
-                        allBlank = false;
-                        break;
-                    }
-                }
-                if (allBlank)
-                    continue;
-                // Integer recordId  is a Priamry key and auto generated by the database.
-                Integer studentId = Integer.parseInt(line[0].trim());
-                String academicYear = line[1].trim();
-                Integer grade = Integer.parseInt(line[2].trim());
-                String subject = line[3].trim();
-                String section = line[4].trim();
-                double q1 = Double.parseDouble(line[5].trim());
-                double q2 = Double.parseDouble(line[6].trim());
-                double q3 = Double.parseDouble(line[7].trim());
-                double q4 = Double.parseDouble(line[8].trim());
-                Integer updatedBy = Integer.parseInt(line[9].trim());
-                LocalDate updateDate = LocalDate.parse(line[10].trim());
-                String comment = line[11].trim();
-// we used setters here instead of passing to constructor b/c recordid is pk and assigned by dbase.
-                StudentRecord s = new StudentRecord();
-
-                s.setStudentId(studentId);
-                s.setSubject(subject);
-                s.setAcademicYear(academicYear);
-                s.setGrade(grade);
-                s.setSection(section);
-                s.setQ1(q1);
-                s.setQ2(q2);
-                s.setQ3(q3);
-                s.setQ4(q4);
-                s.setUpdatedBy(updatedBy);
-                s.setUpdateDate(updateDate);
-                s.setComment(comment);
-                System.out.println(" & s :" + s);
-                studentRecordRepository.save(s);
-            }
-        } catch (NumberFormatException nfe) {
-            nfe.printStackTrace();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    public void loadCsvStudentDataFile() {
-        System.out.println("loadcsvdata method is called ");
-        try {  // 6 - 7 am 1800 -> { 450 }
-            // ClassPathResource refers that everything under resource files will be included in the jar file of this project
-            CSVReader reader = new CSVReader(new InputStreamReader(new ClassPathResource("studentdata.csv").getInputStream()));
-            //TODO new File("C:/path/to/studentdata.csv"); if we need to upload from other location
-            String[] line;
-            boolean isHeader = true;
-            while ((line = reader.readNext()) != null) {
-
-                if (isHeader) {
-                    isHeader = false;
-                    continue;
-                }
-                // Optionally, also skip lines where all fields are blank
-                boolean allBlank = true;
-                for (String field : line) {
-                    if (field != null && !field.trim().isEmpty()) {
-                        allBlank = false;
-                        break;
-                    }
-                }
-                if (allBlank)
-                    continue;
-                Integer studentId = Integer.parseInt(line[0]);                      // 1
-                //System.out.println("1 studentid " + studentId);
-                String firstName = line[1].trim();                                  // 2
-                String middleName = line[2].trim();                                 // 3
-                String lastName = line[3].trim();                                   // 4
-                // String rawdate = line[4].trim();
-                /*  DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");*/
-                LocalDate dateOfBirth = LocalDate.parse(line[4].trim());            // 5
-                String gender = line[5].trim();                                     // 6
-                // String row2 = line[6].trim();
-                LocalDate registrationDate = LocalDate.parse(line[6].trim());       // 7
-                byte[] photo = line[7].getBytes(StandardCharsets.UTF_8);            // 8
-                String kifleKetema = line[8];                                       // 9
-                String kebele = line[9];                                            // 10
-                String houseNumber = line[10];                                      // 11
-                String phone = line[11];                                            // 12
-                String comment = line[12];                                          // 13
-                boolean isActive = Boolean.parseBoolean(line[13]);                  // 14
-                Student student = new Student(studentId, firstName, middleName, lastName, dateOfBirth, gender, registrationDate, photo, kifleKetema, kebele, houseNumber, phone, comment, isActive);
-                System.out.println("ID : " + student.getStudentId() + " & " + student);
-                studentRepository.save(student);
-            }
-        } catch (NumberFormatException nfe) {
-            // System.out.println("Invalid number format on line: " + Arrays.toString(line));
-            nfe.printStackTrace();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    public void loadJSONStudentDataFile() throws IOException {
-        // TODO - currently not functional but if needed to laod data from Json format instead of a csv
-        // https://csvjson.com/csv2json is where we can convert csv to json
-        ObjectMapper mapper = new ObjectMapper();
-        // InputStream inputStream = new FileInputStream("studentdata.json");
-        InputStream inputStream = getClass().getClassLoader().getResourceAsStream("studentdata.json");
-        if (inputStream == null) {
-            throw new FileNotFoundException("File not found in resources folder");
-        }
-        // BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream));
-        TypeReference<List<Student>> typeRef = new TypeReference<>() {};
-        List<Student> people = mapper.readValue(inputStream, typeRef);
-        studentRepository.saveAll(people);
-    }
-}
-
-
-
